Split the storage of llm cache into blob and file.
