+++ modules/llm-cache/storage/file_storage.h
@@ -1,0 +1,26 @@
+/** Copyright 2020-2023 Alibaba Group Holding Limited.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+#ifndef MODULES_LLM_CACHE_STORAGE_FILE_STORAGE_H_
+#define MODULES_LLM_CACHE_STORAGE_FILE_STORAGE_H_
+
+#include "llm-cache/storage/storage.h"
+
+namespace vineyard {
+
+class FileStorage : public IStorage {};
+
+}  // namespace vineyard
+#endif  // MODULES_LLM_CACHE_STORAGE_FILE_STORAGE_H_
